+++ NEWS.md
@@ -1,6 +1,7 @@
 # medfateland 2.5.2
 - information on soil domain and transpiration mode of initialized landscapes
 - overland routing available in new function overland_routing()
+- overland flow allowed in completely flat areas (according to processing order)
 
 # medfateland 2.5.1
 - id mapping existence checked in add_forests